+++ swagger_parser/bin/swagger_parser.dart
@@ -1,6 +1,7 @@
 import 'package:swagger_parser/src/config/yaml_config.dart';
 import 'package:swagger_parser/src/generator/models/generation_statistics.dart';
-import 'package:swagger_parser/src/utils/utils.dart';
+import 'package:swagger_parser/src/utils/output/io_output.dart'
+    if (dart.library.html) 'package:swagger_parser/src/utils/output/web_output.dart';
 import 'package:swagger_parser/swagger_parser.dart';
 
 /// Used for run `dart run swagger_parser`
@@ -40,7 +41,7 @@ Future<void> main(List<String> arguments) async {
       );
     }
 
-    doneMessage(
+    successMessage(
       successSchemasCount: successSchemasCount,
       schemasCount: configs.length,
     );
