+++ swagger_parser/example/.gitignore
@@ -1,3 +1,3 @@
 lib/**
 
-schemas/debug.**
+specefications/debug.**
