+++ swagger_parser/example/swagger_parser.yaml
@@ -2,7 +2,7 @@ swagger_parser:
   # You must provide the file path and/or url to the OpenApi schema.
   
   # Sets the OpenApi schema path directory for api definition.
-  # schema_path: schemas/openapi.json
+  # schema_path: schemes/openapi.json
   # Sets the url of the OpenApi schema.
   # schema_url: https://petstore.swagger.io/v2/swagger.json
